requirements.txt
streamlit
streamlit-option-menu
google-generativeai
PyPDF2
python-docx
python-dotenv
jinja2